sh: Revert __xdiv64_32 size change.
It's only __div64_32 that needs the fix, __xdiv64_32 behaves as
expected with the original size.

+++ arch/sh/lib/div64-generic.c
@@ -5,12 +5,12 @@ FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
 
 #include <linux/types.h>
 
-extern uint32_t __xdiv64_32(u64 n, u32 d);
+extern uint64_t __xdiv64_32(u64 n, u32 d);
 
 uint32_t __div64_32(u64 *xp, u32 y)
 {
 	uint32_t rem;
-	uint32_t q = __xdiv64_32(*xp, y);
+	uint64_t q = __xdiv64_32(*xp, y);
 
 	rem = *xp - q * y;
 	*xp = q;
